remove utils folder, move to lib folder
