+++ .env
@@ -1,0 +1,2 @@
+DB_DIRECT_URL="postgresql://${DB_USER}:${DB_PASSWORD}@${DB_ENDPOINT_ID}.${DB_HOSTNAME}/${DB_NAME}?sslmode=require"
+DB_URL="postgresql://${DB_USER}:${DB_PASSWORD}@${DB_ENDPOINT_ID}-pooler.${DB_HOSTNAME}/${DB_NAME}?sslmode=require&pgbouncer=true"
