+++ .husky/pre-commit
@@ -1,12 +1,1 @@
-#!/usr/bin/env sh
-
-echo "Running linting checks..."
 yarn run lint
-
-# Check if linting or code formatting checks failed
-if [ $? -ne 0 ]; then
-  echo "Linting failed. Aborting commit."
-  exit 1
-fi
-
-exit 0
