+++ README.md
@@ -1,34 +1,14 @@
-This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
+# Setup
 
-## Getting Started
+We rely on Neon's postgresql DB + Prisma ORM client.
 
-First, run the development server:
+To set this up, make sure the environment variables are set. For local testing, we can rely on a local psql connection.
 
-```bash
-npm run dev
-# or
-yarn dev
-# or
-pnpm dev
-```
+Reset the DB:
+`yarn run db:reset`
 
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+Migrate the DB + seed it.
+`yarn run db:migrate`
 
-You can start editing the page by modifying `app/page.tsx`. The page auto-updates as you edit the file.
 
-This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
-
-## Learn More
-
-To learn more about Next.js, take a look at the following resources:
-
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
-
-You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
-
-## Deploy on Vercel
-
-The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
-
-Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
+If any changes are made to the schema / migrations, we can run `yarn run db:migrate` again.
