+++ app/audits/page.tsx
@@ -1,10 +1,10 @@
 import { redirect } from "next/navigation";
 
 import Audit, { AuditHolder } from "@/components/pages/Audits";
-import { audits } from "@/utils/constants";
+import { audits } from "@/lib/constants";
 import { Section } from "@/components/Common";
 import { H2 } from "@/components/Text";
-import { AuditSSRI } from "@/utils/types";
+import { AuditSSRI } from "@/lib/types";
 
 const getData = (status: string): AuditSSRI => {
   if (!["open", "pending", "closed"].includes(status)) {
