+++ components/Nav/index.tsx
@@ -5,7 +5,7 @@ import { useRef, useReducer } from "react";
 import { usePathname } from "next/navigation";
 
 import { Arrow, Twitter, Discord, Github } from "@/assets";
-import { navItems } from "@/utils/constants";
+import { navItems } from "@/lib/constants";
 import { Span } from "@/components/Text";
 import SmartLink from "@/components/Link";
 import { LogoIcon } from "@/components/Icon";
