+++ components/Nav/web3.tsx
@@ -11,7 +11,7 @@ import { JazziconClient, IconMedium, IconSmall } from "@/components/Icon";
 import { ButtonLight } from "@/components/Button";
 import { DropDown } from "@/components/Tooltip";
 // import { Loader } from "@/components/Common";
-import { ChainPresets } from "@/utils/constants/chains";
+import { ChainPresets } from "@/lib/constants/chains";
 
 export default (): JSX.Element => {
   const { open } = useWeb3Modal();
