+++ package.json
@@ -9,9 +9,19 @@
     "postinstall": "husky init",
     "lint": "next lint && prettier --check \"{app,assets,components,hooks,providers,theme,utils}/**/*.{ts,tsx,js,jsx}\" && tsc --noEmit",
     "format": "prettier --write \"{app,assets,components,hooks,providers,theme,utils}/**/*.{ts,tsx,js,jsx}\"",
-    "prepare": "husky"
+    "prepare": "husky",
+    "db:studio": "prisma studio",
+    "db:migrate": "prisma migrate dev",
+    "db:seed": "prisma db seed",
+    "db:reset": "prisma migrate reset"
+  },
+  "prisma": {
+    "seed": "tsx prisma/seed.ts"
   },
   "dependencies": {
+    "@neondatabase/serverless": "^0.8.1",
+    "@prisma/adapter-neon": "^5.9.1",
+    "@prisma/client": "^5.9.1",
     "@types/node": "20.2.5",
     "@types/react": "18.2.9",
     "@types/react-dom": "18.2.4",
@@ -20,14 +30,15 @@
     "@vercel/speed-insights": "^1.0.10",
     "@web3modal/ethereum": "^2.7.0",
     "@web3modal/react": "^2.7.0",
-    "drizzle-orm": "^0.29.3",
     "encoding": "^0.1.13",
     "eslint": "8.42.0",
     "eslint-config-next": "13.4.4",
     "gray-matter": "^4.0.3",
     "lokijs": "^1.5.12",
     "next": "13.4.4",
+    "pg": "^8.11.3",
     "pino-pretty": "^10.0.1",
+    "prisma": "^5.9.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-jazzicon": "^1.0.4",
@@ -37,18 +48,19 @@
     "styled-components": "^6.0.3",
     "typescript": "5.1.3",
     "viem": "^1.3.0",
-    "wagmi": "^1.3.9"
+    "wagmi": "^1.3.9",
+    "ws": "^8.16.0"
   },
   "devDependencies": {
     "@next/eslint-plugin-next": "^13.4.4",
     "@typescript-eslint/eslint-plugin": "^5.59.7",
     "@typescript-eslint/parser": "^5.59.7",
-    "drizzle-kit": "^0.20.14",
     "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-prettier": "^4.2.1",
     "husky": "^9.0.11",
-    "prettier": "^2.8.8"
+    "prettier": "^2.8.8",
+    "tsx": "^4.7.1"
   }
 }
